checkpoint

+++ HandheldCompanion/Devices/Valve/SteamDeck.cs
@@ -215,7 +215,7 @@ public class SteamDeck : IDevice
         int value = SupportedDevice?.BatteryTempLE == true ?
             ((data[1] << 8) + data[0]) :
             ((data[0] << 8) + data[1]);
-        return (float)(value - 0x0AAC) / 10.0f;
+        return (value - 0x0AAC) / 10.0f;
     }
 
     public int? GetMaxBatteryCharge()
